+++ example/react-demo/src/main.tsx
@@ -1,16 +1,11 @@
 import { StrictMode } from 'react'
 import ReactDOM from 'react-dom/client'
-import { PermissionProvider } from '@permission-kit/react'
+import { PermissionProvider } from '@eycraf/permission-kit-react'
 import App from './App'
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <StrictMode>
-    <PermissionProvider
-      permissions={[
-        'user.view',
-        'user.create'
-      ]}
-    >
+    <PermissionProvider permissions={['user.view', 'user.create']}>
       <App />
     </PermissionProvider>
   </StrictMode>