+++ packages/core/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @permission-kit/core
+# @eycraf/permission-kit-core
 
 ## 0.0.1